+++ tests/uitests/src/test/snapshots/images/io.element.android.tests.uitests_ScreenshotTest_preview_tests[io.element.android.libraries.matrix.ui.components_null_DefaultGroup_MatrixUserHeaderDarkPreview__1,NEXUS_5,1,en].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:90da6bebf728f89f42cf51d0cdf438d74f96504dc0758c647a4f07a88ebbcac1
+size 20472
