+++ tests/uitests/src/test/snapshots/images/io.element.android.tests.uitests_ScreenshotTest_preview_tests[io.element.android.libraries.matrix.ui.components_null_DefaultGroup_MatrixUserRowLightPreview_,NEXUS_5,1,en].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:33a5cbc319ae522a937ca19543dfce68e8971a0a1ff82ed3e7991d68eda82d5d
+size 12666
